working buttons admin dashboard

+++ src/components/admin/DeleteProduct.js
@@ -1,43 +1,49 @@
 import { Button, ButtonGroup, ToggleButton } from 'react-bootstrap';
 import Swal from 'sweetalert2';
 
-
-
 export default function DeleteOrder(prop) {
 	
 	let productId = prop.prodId;
-
 	
-
 	const deleteProduct = () => {
 
-		fetch('http://localhost:4000/products/delete', {
-			method: 'DELETE',
-			headers: {
-				'Content-Type': 'application/json',
-				Authorization: `Bearer ${ localStorage.getItem('accessToken')}`
-			},
-			body: JSON.stringify({
-				productId: productId
-			})
-		}).then(
-			Swal.fire({
-			    title: "Product Successfully Deleted",
-			    text: "Click OK to Exit",
-			    type: "success",
-			    icon: 'success'
-			}).then(function() {
-			    window.location.reload(false);
-			})
-		)
-		
-				
+		Swal.fire({
+		  title: 'Are you sure?',
+		  text: "You won't be able to revert this!",
+		  icon: 'warning',
+		  showCancelButton: true,
+		  confirmButtonColor: '#3085d6',
+		  cancelButtonColor: '#d33',
+		  confirmButtonText: 'Yes, delete it!'
+		})
+		.then((result) => {
+		  	
+		  	if (result.isConfirmed) {
+
+		  		fetch('http://localhost:4000/products/delete', {
+		  			method: 'DELETE',
+		  			headers: {
+		  				'Content-Type': 'application/json',
+		  				Authorization: `Bearer ${ localStorage.getItem('accessToken')}`
+		  			},
+		  			body: JSON.stringify({
+		  				productId: productId
+		  			})
+		  		})			  	
+		  		.then(
+		  			Swal.fire({
+		  				title: 'Item Successfully Deleted',
+		  				text: "Click Ok to Exit",
+		  				icon: 'success',
+		  				confirmButtonColor: '#3085d6',
+		  				confirmButtonText: 'Ok'
+		  			})
+		  			.then(result => window.location.reload(false))
+		  		)	
+			}
+		})
 	}
-
-	
-
-	
-	
+		
 						
 	return(
 		<>
@@ -48,6 +54,5 @@ export default function DeleteOrder(prop) {
 			</div>      
 			
 		</>
-
-		);
+	)
 }
